Move tags column into create table statement
This was originally an ALTER TABLE command to migrate any tables that
used the old version seamlessly onto the new. The ALTER TABLE statement
requires a table lock which conflicted with the transaction locks held
by any kewpie workers that were already running. This will be a breaking
change that will fail on any existing tables that happen to be running
the old schema that have not yet been migrated.