Handle transactions entirely internally
This removes the theoretical ability to pass a transaction into the Pop
and Subscribe methods. This was not being used in practice as there's no
actual use case for making the consumption of a task part of a broader
transaction context.

Now that the transaction is exclusively internal we can properly Commit
it after we poll the queue and find nothing there, leaving the
connection available for our next poll.