Set content-type to application/json for cloud tasks
Turns out Google Cloud Tasks doesn't set this without being explicitly
told. Experimentally, it might also be inconsistent with whether it
leaves it blank or sets it to application/octet-stream.

If it's blank, then r.ParseForm() will consume the body, leaving the
FromHTTP function to read in a blank body and fail. If the Content-Type
header is set to either json or octet-stream, the an r.ParseForm() call
will ignore the body, leaving it intact for FromHTTP.

This is relevant in cases where there may be some middleware in line
before your SubscribeHTTP call that might call ParseForm and gazzump
your body.